# companion/openapi.yaml
openapi: 3.1.0
info:
  title: Companion API
  version: 0.1.0
  description: |
    API del empleado IA central del ecosistema.

    Companion acepta API key o Bearer JWT interno. Cuando hay principal
    autenticado, `org_id` se interpreta como customer org/work context: la
    empresa cliente donde trabaja el empleado IA. Headers como `X-Org-ID` y
    `X-User-ID` quedan como compatibilidad de dev/test, no como fuente canonica
    productiva.
servers:
  - url: http://localhost:18085
security:
  - apiKeyAuth: []
  - bearerAuth: []
paths:
...
  /v1/virtual-employees:
    get:
      summary: Lista Virtual Employees de la customer org
      description: API publica recomendada para trabajadores digitales persistentes. `tenant_id` es UUID publico y Companion mantiene `org_id + product_surface` solo como contexto interno de transicion.
      parameters:
        - in: query
          name: lifecycle
          schema: { type: string, enum: [active, archived, trashed, all], default: active }
      responses:
        "200":
          description: Virtual Employee list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/VirtualEmployee" }
    post:
      summary: Crea un Virtual Employee
      description: Crea un trabajador digital persistente con ID automatico. No acepta `agent_id`, `metadata`, `job_title`, `mission`, `responsibilities`, `allowed_tools`, `memory_enabled` ni `memory_scope_id`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VirtualEmployeeInput" }
      responses:
        "201":
          description: Virtual Employee created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
  /v1/employee-profiles:
    get:
      summary: Lista Employee Profiles
      description: |
        Superficie publica recomendada para perfiles de Virtual Employees.
        En v1 reutiliza Agent Profiles internamente; `profile_id` es el UUID
        publico del perfil y `profile_key` conserva la key tecnica usada por
        Agent/Runtime durante la transicion.
      parameters:
        - in: query
          name: lifecycle
          schema: { type: string, enum: [active, archived, trash, all], default: active }
        - in: query
          name: include_archived
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Employee Profile list
          content:
            application/json:
              schema:
                type: object
                properties:
                  employee_profiles:
                    type: array
                    items: { $ref: "#/components/schemas/EmployeeProfile" }
                  profiles:
                    type: array
                    items: { $ref: "#/components/schemas/EmployeeProfile" }
    post:
      summary: Crea un Employee Profile
      description: Crea un perfil con UUID publico automatico; el cliente puede enviar `profile_key` humano o dejarlo autogenerado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeProfileInput" }
      responses:
        "201":
          description: Employee Profile created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeProfile" }
...
  /v1/virtual-employees/{employee_id}:
    get:
      summary: Obtiene un Virtual Employee
      description: API publica recomendada. `employee_id` es UUID publico del Virtual Employee.
      parameters:
        - in: path
          name: employee_id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Virtual Employee
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
    patch:
      summary: Actualiza un Virtual Employee
      description: Actualiza solo el core del Employee. No cambia status; para lifecycle usar `/status`.
      parameters:
        - in: path
          name: employee_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VirtualEmployeeInput" }
      responses:
        "200":
          description: Virtual Employee updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
  /v1/virtual-employees/{employee_id}/status:
    post:
      summary: Cambia el status de un Virtual Employee
      parameters:
        - in: path
          name: employee_id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/EmployeeStatus" }
      responses:
        "200":
          description: Virtual Employee status changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VirtualEmployee" }
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    VirtualEmployee:
      type: object
      description: Trabajador digital persistente con identidad propia. No es un Agent ni un alias de modulo tecnico de agents.
      required: [employee_id, tenant_id, name, supervisor_user_id, status, job_role_id, profile_id, autonomy, capability_ids]
      properties:
        employee_id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid }
        name: { type: string }
        supervisor_user_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/EmployeeStatus" }
        job_role_id: { type: string, format: uuid }
        profile_id: { type: string, format: uuid }
        autonomy: { $ref: "#/components/schemas/AutonomyLevel" }
        capability_ids:
          type: array
          items: { type: string, format: uuid }
        memory_id: { type: string, format: uuid, nullable: true }
    VirtualEmployeeInput:
      type: object
      required: [name, supervisor_user_id, job_role_id, profile_id, autonomy]
      additionalProperties: false
      properties:
        name: { type: string }
        supervisor_user_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/EmployeeStatus" }
        job_role_id: { type: string, format: uuid }
        profile_id: { type: string, format: uuid }
        autonomy: { $ref: "#/components/schemas/AutonomyLevel" }
        capability_ids:
          type: array
          items: { type: string, format: uuid }
        memory_id: { type: string, format: uuid, nullable: true }
    EmployeeStatus:
      type: string
      enum: [draft, active, disabled, suspended, archived, trashed, error]
    AutonomyLevel:
      type: string
      enum: [A0, A1, A2, A3, A4, A5]
    EmployeeProfileInput:
      type: object
      required: [name, system_prompt]
      properties:
        profile_key:
          type: string
          description: Human-readable technical key. Optional on create; generated from name/family/version when omitted.
        family_id: { type: string }
        version_label: { type: string, default: v1 }
        name: { type: string }
        description: { type: string }
        system_prompt: { type: string }
        max_autonomy: { type: string, enum: [A0, A1, A2, A3, A4, A5], default: A1 }
        default_capability_ids:
          type: array
          description: Target model field. During transition it maps to the existing profile capability allowlist.
          items: { type: string, format: uuid }
        allowed_capabilities:
          type: array
          deprecated: true
          description: Legacy capability keys used by the technical runtime profile.
          items: { type: string }
        allowed_tools:
          type: array
          deprecated: true
          description: Legacy direct tool allowlist. Employees should reference capabilities instead.
          items: { type: string }
        memory_policy: { type: object, additionalProperties: true }
        llm_config: { type: object, additionalProperties: true }
        enabled: { type: boolean, default: true }
    EmployeeProfile:
      allOf:
        - $ref: "#/components/schemas/EmployeeProfileInput"
        - type: object
          required: [profile_id, profile_key, name, system_prompt, max_autonomy, status]
          properties:
            id: { type: string, format: uuid }
            profile_id:
              type: string
              format: uuid
              description: Public EmployeeProfile UUID.
            profile_key:
              type: string
              description: Compatibility key currently used by the technical runtime profile.
            status: { type: string, enum: [active, disabled, archived, trashed] }
            archived_at: { type: string, format: date-time }
            trashed_at: { type: string, format: date-time }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...